+++ TowerForge/TowerForge/Metrics/Achievements/AchievementStatsLinkDatabase.swift
@@ -9,5 +9,5 @@ import Foundation
 
 /// All achievements must be linked to a specific statistic
 class AchievementStatsLinkDatabase {
-    var achievementLinks: [TFAchievementTypeWrapper: [Statistic]] = [:]
+    var achievementLinks: [AchievementTypeWrapper: [Statistic]] = [:]
 }
