+++ TowerForge/TowerForge/Metrics/Achievements/AchievementTypeWrapper.swift
@@ -1,0 +1,20 @@
+//
+//  AchievementTypeWrapper.swift
+//  TowerForge
+//
+//  Created by Rubesh on 14/4/24.
+//
+
+import Foundation
+
+struct AchievementTypeWrapper: Equatable, Hashable {
+    let type: Achievement.Type
+
+    static func == (lhs: Self, rhs: Self) -> Bool {
+        lhs.type == rhs.type
+    }
+
+    func hash(into hasher: inout Hasher) {
+        hasher.combine(ObjectIdentifier(type))
+    }
+}
