+++ TowerForge/TowerForge/Metrics/Statistics/Implemented/DefaultStatistic.swift
