+++ TowerForge/TowerForge/Metrics/Statistics/Implemented/TotalDeathsStatistic.swift
