+++ TowerForge/TowerForge/Metrics/Statistics/Implemented/TotalGamesStatistic.swift
