+++ TowerForge/TowerForge/Metrics/Statistics/Implemented/TotalKillsStatistic.swift
